refactor entity
